binder_prio: Fix: Use sysfs to detect MIUI/HyperOS ROM
This commit addresses the build issue caused by the use of the
internal kern_path() function.  kern_path() is not a stable ABI
and can change between kernel versions, leading to build failures
for vendor modules.

This commit replaces the kern_path() call with a check for the
ro.miui.ui.version.name sysfs property.  This property is a more
reliable way to detect if the system is running on a MIUI/HyperOS
ROM.

This change makes the binder_prio module more robust and avoids
reliance on internal kernel functions.

Signed-off-by: klozz <carlosj@klozz.dev>